serialized/deserialized user, added config folder with passport-config file, deleted eslint stuff
