+++ client/package.json
@@ -24,8 +24,5 @@
     "not ie <= 11",
     "not op_mini all"
   ],
-  "devDependencies": {
-    "eslint": "^7.1.0",
-    "eslint-plugin-react": "^7.20.0"
-  }
+  "devDependencies": {}
 }
