+++ package.json
@@ -1,37 +1,34 @@
 {
-  "name": "presume",
-  "version": "1.0.0",
-  "description": "",
-  "main": "server.js",
-  "scripts": {
-    "start": "if-env NODE_ENV=production && npm run start:prod || npm run start:dev",
-    "start:prod": "node server.js",
-    "start:dev": "concurrently \"nodemon --ignore 'client/*'\" \"npm run client\"",
-    "client": "cd client && npm run start",
-    "install": "cd client && npm install",
-    "build": "cd client && npm run build",
-    "heroku-postbuild": "npm run build",
-    "seed": "node scripts/seedDB.js",
-    "test": "npm run lint",
-    "lint": "concurrently \"eslint --quiet .\" \"cd client && npm run lint\"",
-    "fix": "concurrently \"eslint --fix . \" \"cd client && npm run lint\""
-  },
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "concurrently": "^4.1.0",
-    "eslint": "^6.8.0",
-    "eslint-config-prettier": "^6.11.0",
-    "eslint-plugin-prettier": "^3.1.3",
-    "nodemon": "^1.18.7",
-    "prettier": "^2.0.5"
-  },
-  "dependencies": {
-    "dotenv": "^8.2.0",
-    "express": "^4.16.3",
-    "if-env": "^1.0.4",
-    "mongoose": "^5.8.11",
-    "passport": "^0.4.1",
-    "passport-google-oauth": "^2.0.0"
-  }
+	"name": "presume",
+	"version": "1.0.0",
+	"description": "",
+	"main": "server.js",
+	"scripts": {
+		"start": "if-env NODE_ENV=production && npm run start:prod || npm run start:dev",
+		"start:prod": "node server.js",
+		"start:dev": "concurrently \"nodemon --ignore 'client/*'\" \"npm run client\"",
+		"client": "cd client && npm run start",
+		"install": "cd client && npm install",
+		"build": "cd client && npm run build",
+		"heroku-postbuild": "npm run build",
+		"seed": "node scripts/seedDB.js",
+		"test": "npm run lint",
+		"lint": "concurrently \"eslint --quiet .\" \"cd client && npm run lint\"",
+		"fix": "concurrently \"eslint --fix . \" \"cd client && npm run lint\""
+	},
+	"author": "",
+	"license": "ISC",
+	"devDependencies": {
+		"concurrently": "^4.1.0",
+		"nodemon": "^1.18.7",
+		"prettier": "^2.0.5"
+	},
+	"dependencies": {
+		"dotenv": "^8.2.0",
+		"express": "^4.16.3",
+		"if-env": "^1.0.4",
+		"mongoose": "^5.8.11",
+		"passport": "^0.4.1",
+		"passport-google-oauth": "^2.0.0"
+	}
 }
