+++ routes/htmlRoutes.js
@@ -7,13 +7,5 @@ function authCheck(req, res, next) {
 router.get('/', authCheck, (req, res) => {
 	res.send('you have successfully logged in!');
 });
-// authenticating users
-router.get('/login/success', (req, res) => {
-	res.send('you have successfully logged in!');
-});
-router.get('/login/failure', (req, res) => {
-	res.send('you have failed to log in');
-});
-router.get('/logout');
 
 module.exports = router;
